pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "verifybot"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^2.0.1"
python = "^3.8"
[tool.poetry.dependencies."discord.py"]
branch = "modals"
git = "https://github.com/Chiggy-Playz/enhanced-discord.py"
[tool.poetry.dev-dependencies]
